Add files via upload
Buying computer parts. 
The valid choices are displayed to the user. 
The user selects the computer parts he wishes to purchase, and the parts are added to his shopping cart list.
If the user wishes to remove an item from the list, he can simply select the item again as all duplicates in the shopping cart list are removed.
If the user is finished adding items, he must press 0 to exit.