Description of program
This program was created to practice working with tuples in Python.
In this program the user selects music to be played from a jukebox.
The user selects the album, then song from the album. 
The program then continues to loop until the user is finished using the jukebox.
The user selects the album and song by typing in the number associated with the album or song.
In order to exit the program the user must simply select an invalid input when selecting an album.